refactor: 🔨hex to UIColor method UIColor extension 으로 파일 분리(#8)
