+++ ios/IssueTracker/IssueTracker/IssueList/LabelView.swift
@@ -33,35 +33,16 @@ extension LabelView {
     }
     
     func fillUI(with label: Label) {
-        self.text = label.name
-        self.backgroundColor = hexStringToUIColor(hex: label.color)
-        print(intrinsicContentSize.width)
+        text = label.name
+        backgroundColor = UIColor.hexToUIColor(hex: label.color)
+        configureConstraint()
+    }
+    
+    private func configureConstraint() {
         NSLayoutConstraint.activate([
             heightAnchor.constraint(equalToConstant: intrinsicContentSize.height + 10),
             widthAnchor.constraint(greaterThanOrEqualToConstant: intrinsicContentSize.width + 20)
         ])
-    }
-    
-    func hexStringToUIColor (hex:String) -> UIColor {
-        var cString:String = hex.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
-
-        if (cString.hasPrefix("#")) {
-            cString.remove(at: cString.startIndex)
-        }
-
-        if ((cString.count) != 6) {
-            return UIColor.gray
-        }
-
-        var rgbValue:UInt64 = 0
-        Scanner(string: cString).scanHexInt64(&rgbValue)
-
-        return UIColor(
-            red: CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0,
-            green: CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0,
-            blue: CGFloat(rgbValue & 0x0000FF) / 255.0,
-            alpha: CGFloat(1.0)
-        )
     }
     
 }
